Removed auth.js and added keep alive code
