+++ app.js
@@ -14,7 +14,4 @@ app.use(morgan('dev'));
 var UserController = require('./user/UserController');
 app.use('/api/', UserController);
 
-var AuthController = require('./auth/AuthController');
-app.use('/api/auth', AuthController);
-
 module.exports = app;
