+++ user/UserController.js
@@ -8,18 +8,27 @@ router.use(bodyParser.urlencoded({ extended: true }));
 router.use(bodyParser.json());
 
 var User = require('./User');
+var bcrypt = require('bcryptjs');
 
 // CREATES A NEW USER
-router.post('/', function (req, res) {
-    User.create({
-        name : req.body.name,
-        email : req.body.email,
-        password : req.body.password
-    },
-    function (err, user) {
-        if (err) return res.status(500).send("There was a problem adding the information to the database.");
-        res.status(200).send(user);
+router.post('/register', function(req, res) {
+
+  var hashedPassword = bcrypt.hashSync(req.body.password, 8);
+  console.log(`name: ${req.body.name} email: ${req.body.email} password: ${req.body.password}`);
+
+  User.create({
+    name : req.body.name,
+    email : req.body.email,
+    password : hashedPassword
+  },
+  function (err, user) {
+    if (err) return res.status(500).send("There was a problem registering the user.")
+    // create a token
+    var token = jwt.sign({ id: user._id }, config.secret, {
+      expiresIn: 86400 // expires in 24 hours
     });
+    res.status(200).send({ auth: true, token: token });
+  });
 });
 
 // Welcome users
@@ -51,6 +60,7 @@ router.get('/users', function (req, res) {
 
 // route to authenticate a user (POST http://localhost:3000/api/authenticate)
 router.post('/authenticate', function (req, res) {
+  
   // find the user
   User.findOne({
     name: req.body.name
